Handle case where the root project's child projects have subprojects
Previously the plugin would only evaluate the `rootProject`'s children
projects.  This works fine if the root's children are the actual projects
in the case where the root's children are only a root for rootProjects's
children's children, rootProject's grand children, this is not enough.

Now for each child project of root we check for subproject of the children.

If grand children project's are found then add their dependencies to the
report