added more checks for the debug logging and restrict configuration
Added code to check if debug logging is on to prevent proccessing of the
dependency tree log messages.  Also check the configuration on the child
dependencies and make sure that it matches the configuration we are
interested in.  The getChildren method for a resolved dependency appears
to get the resolved dependencies for all configurations.  So we need to
restrict it.