Allow the project name and version to be specified in build.gradle file
Make code changes to expose the properties for the hub project name and
version as items that can be declared in the build.gradle file.  The
properties are hubProjectName and hubProjectVersion.  The user is also
able to define them as properties in the gradle.properties file and on
the command line as a -P option.